requirements.txt
openai>=1.58.0
pdf2image>=1.16.0
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
typer>=0.9.0